# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamVault API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Genre {
  id: number;
  name: string;
}

export interface GenreList {
  genres: Genre[];
}

export interface MediaItem {
  id: number;
  title?: string;
  name?: string;
  overview?: string;
  poster_path?: string | null;
  backdrop_path?: string | null;
  vote_average?: number;
  vote_count?: number;
  release_date?: string;
  first_air_date?: string;
  media_type?: string;
  genre_ids?: number[];
  popularity?: number;
}

export interface MediaPage {
  page: number;
  results: MediaItem[];
  total_pages: number;
  total_results: number;
}

export interface CastMember {
  id: number;
  name: string;
  character?: string;
  profile_path?: string | null;
  order?: number;
}

export interface Video {
  id: string;
  key: string;
  name: string;
  site: string;
  type: string;
}

export type MovieDetailCredits = {
  cast?: CastMember[];
};

export type MovieDetailVideos = {
  results?: Video[];
};

export interface MovieDetail {
  id: number;
  title: string;
  overview?: string;
  poster_path?: string | null;
  backdrop_path?: string | null;
  vote_average?: number;
  vote_count?: number;
  release_date?: string;
  runtime?: number | null;
  genres?: Genre[];
  tagline?: string | null;
  status?: string;
  credits?: MovieDetailCredits;
  videos?: MovieDetailVideos;
  similar?: MediaPage;
}

export interface Episode {
  id: number;
  episode_number: number;
  season_number: number;
  name: string;
  overview?: string;
  still_path?: string | null;
  air_date?: string;
  vote_average?: number;
  runtime?: number | null;
}

export interface Season {
  id: number;
  season_number: number;
  name: string;
  overview?: string;
  poster_path?: string | null;
  episode_count?: number;
  air_date?: string;
}

export interface SeasonDetail {
  id: number;
  season_number: number;
  name: string;
  overview?: string;
  poster_path?: string | null;
  episodes: Episode[];
}

export type TvDetailCredits = {
  cast?: CastMember[];
};

export type TvDetailVideos = {
  results?: Video[];
};

export interface TvDetail {
  id: number;
  name: string;
  overview?: string;
  poster_path?: string | null;
  backdrop_path?: string | null;
  vote_average?: number;
  vote_count?: number;
  first_air_date?: string;
  last_air_date?: string;
  number_of_seasons?: number;
  number_of_episodes?: number;
  status?: string;
  tagline?: string | null;
  genres?: Genre[];
  seasons?: Season[];
  credits?: TvDetailCredits;
  videos?: TvDetailVideos;
  similar?: MediaPage;
}

export type GetTrendingParams = {
  type?: GetTrendingType;
  timeWindow?: GetTrendingTimeWindow;
  page?: number;
};

export type GetTrendingType =
  (typeof GetTrendingType)[keyof typeof GetTrendingType];

export const GetTrendingType = {
  all: "all",
  movie: "movie",
  tv: "tv",
} as const;

export type GetTrendingTimeWindow =
  (typeof GetTrendingTimeWindow)[keyof typeof GetTrendingTimeWindow];

export const GetTrendingTimeWindow = {
  day: "day",
  week: "week",
} as const;

export type SearchContentParams = {
  query: string;
  type?: SearchContentType;
  page?: number;
};

export type SearchContentType =
  (typeof SearchContentType)[keyof typeof SearchContentType];

export const SearchContentType = {
  multi: "multi",
  movie: "movie",
  tv: "tv",
} as const;

export type GetPopularMoviesParams = {
  page?: number;
};

export type GetTopRatedMoviesParams = {
  page?: number;
};

export type GetPopularTvParams = {
  page?: number;
};

export type GetTopRatedTvParams = {
  page?: number;
};

export type DiscoverMoviesParams = {
  genreId?: number;
  sortBy?: string;
  page?: number;
};

export type DiscoverTvParams = {
  genreId?: number;
  sortBy?: string;
  page?: number;
};