# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreamVault API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: trending
    description: Trending content
  - name: search
    description: Search content
  - name: movies
    description: Movie operations
  - name: tv
    description: TV show operations
  - name: genres
    description: Genre operations
  - name: discover
    description: Discover content
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /trending:
    get:
      operationId: getTrending
      tags: [trending]
      summary: Get trending movies and TV shows
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [all, movie, tv]
            default: all
        - name: timeWindow
          in: query
          schema:
            type: string
            enum: [day, week]
            default: week
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Trending content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPage"

  /search:
    get:
      operationId: searchContent
      tags: [search]
      summary: Search for movies and TV shows
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [multi, movie, tv]
            default: multi
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPage"

  /movie/{id}:
    get:
      operationId: getMovie
      tags: [movies]
      summary: Get movie details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movie details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDetail"

  /tv/{id}:
    get:
      operationId: getTvShow
      tags: [tv]
      summary: Get TV show details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: TV show details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TvDetail"

  /tv/{id}/season/{seasonNumber}:
    get:
      operationId: getTvSeason
      tags: [tv]
      summary: Get TV season details with episodes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: seasonNumber
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Season details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeasonDetail"

  /movies/popular:
    get:
      operationId: getPopularMovies
      tags: [movies]
      summary: Get popular movies
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Popular movies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPage"

  /movies/top-rated:
    get:
      operationId: getTopRatedMovies
      tags: [movies]
      summary: Get top rated movies
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Top rated movies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPage"

  /tv/popular:
    get:
      operationId: getPopularTv
      tags: [tv]
      summary: Get popular TV shows
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Popular TV shows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPage"

  /tv/top-rated:
    get:
      operationId: getTopRatedTv
      tags: [tv]
      summary: Get top-rated TV shows
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Top-rated TV shows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPage"

  /genres/movies:
    get:
      operationId: getMovieGenres
      tags: [genres]
      summary: Get list of movie genres
      responses:
        "200":
          description: Movie genres
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenreList"

  /genres/tv:
    get:
      operationId: getTvGenres
      tags: [genres]
      summary: Get list of TV genres
      responses:
        "200":
          description: TV genres
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenreList"

  /discover/movies:
    get:
      operationId: discoverMovies
      tags: [discover]
      summary: Discover movies by genre and sort
      parameters:
        - name: genreId
          in: query
          schema:
            type: integer
        - name: sortBy
          in: query
          schema:
            type: string
            default: popularity.desc
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Discovered movies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPage"

  /discover/tv:
    get:
      operationId: discoverTv
      tags: [discover]
      summary: Discover TV shows by genre and sort
      parameters:
        - name: genreId
          in: query
          schema:
            type: integer
        - name: sortBy
          in: query
          schema:
            type: string
            default: popularity.desc
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Discovered TV shows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPage"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Genre:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name

    GenreList:
      type: object
      properties:
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
      required:
        - genres

    MediaItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        name:
          type: string
        overview:
          type: string
        poster_path:
          type: string
          nullable: true
        backdrop_path:
          type: string
          nullable: true
        vote_average:
          type: number
        vote_count:
          type: integer
        release_date:
          type: string
        first_air_date:
          type: string
        media_type:
          type: string
        genre_ids:
          type: array
          items:
            type: integer
        popularity:
          type: number
      required:
        - id

    MediaPage:
      type: object
      properties:
        page:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"
        total_pages:
          type: integer
        total_results:
          type: integer
      required:
        - page
        - results
        - total_pages
        - total_results

    CastMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        character:
          type: string
        profile_path:
          type: string
          nullable: true
        order:
          type: integer
      required:
        - id
        - name

    Video:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        name:
          type: string
        site:
          type: string
        type:
          type: string
      required:
        - id
        - key
        - name
        - site
        - type

    MovieDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        overview:
          type: string
        poster_path:
          type: string
          nullable: true
        backdrop_path:
          type: string
          nullable: true
        vote_average:
          type: number
        vote_count:
          type: integer
        release_date:
          type: string
        runtime:
          type: integer
          nullable: true
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
        tagline:
          type: string
          nullable: true
        status:
          type: string
        credits:
          type: object
          properties:
            cast:
              type: array
              items:
                $ref: "#/components/schemas/CastMember"
        videos:
          type: object
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/Video"
        similar:
          $ref: "#/components/schemas/MediaPage"
      required:
        - id
        - title

    Episode:
      type: object
      properties:
        id:
          type: integer
        episode_number:
          type: integer
        season_number:
          type: integer
        name:
          type: string
        overview:
          type: string
        still_path:
          type: string
          nullable: true
        air_date:
          type: string
        vote_average:
          type: number
        runtime:
          type: integer
          nullable: true
      required:
        - id
        - episode_number
        - season_number
        - name

    Season:
      type: object
      properties:
        id:
          type: integer
        season_number:
          type: integer
        name:
          type: string
        overview:
          type: string
        poster_path:
          type: string
          nullable: true
        episode_count:
          type: integer
        air_date:
          type: string
      required:
        - id
        - season_number
        - name

    SeasonDetail:
      type: object
      properties:
        id:
          type: integer
        season_number:
          type: integer
        name:
          type: string
        overview:
          type: string
        poster_path:
          type: string
          nullable: true
        episodes:
          type: array
          items:
            $ref: "#/components/schemas/Episode"
      required:
        - id
        - season_number
        - name
        - episodes

    TvDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        overview:
          type: string
        poster_path:
          type: string
          nullable: true
        backdrop_path:
          type: string
          nullable: true
        vote_average:
          type: number
        vote_count:
          type: integer
        first_air_date:
          type: string
        last_air_date:
          type: string
        number_of_seasons:
          type: integer
        number_of_episodes:
          type: integer
        status:
          type: string
        tagline:
          type: string
          nullable: true
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
        seasons:
          type: array
          items:
            $ref: "#/components/schemas/Season"
        credits:
          type: object
          properties:
            cast:
              type: array
              items:
                $ref: "#/components/schemas/CastMember"
        videos:
          type: object
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/Video"
        similar:
          $ref: "#/components/schemas/MediaPage"
      required:
        - id
        - name