# lib/api-zod/src/generated/types/mediaPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamVault API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MediaItem } from "./mediaItem";

export interface MediaPage {
  page: number;
  results: MediaItem[];
  total_pages: number;
  total_results: number;
}