# lib/api-zod/src/generated/types/movieDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamVault API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Genre } from "./genre";
import type { MediaPage } from "./mediaPage";
import type { MovieDetailCredits } from "./movieDetailCredits";
import type { MovieDetailVideos } from "./movieDetailVideos";

export interface MovieDetail {
  id: number;
  title: string;
  overview?: string;
  poster_path?: string | null;
  backdrop_path?: string | null;
  vote_average?: number;
  vote_count?: number;
  release_date?: string;
  runtime?: number | null;
  genres?: Genre[];
  tagline?: string | null;
  status?: string;
  credits?: MovieDetailCredits;
  videos?: MovieDetailVideos;
  similar?: MediaPage;
}