# lib/api-zod/src/generated/types/season.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Season {
  id: number;
  season_number: number;
  name: string;
  overview?: string;
  poster_path?: string | null;
  episode_count?: number;
  air_date?: string;
}