# lib/api-zod/src/generated/types/seasonDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamVault API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Episode } from "./episode";

export interface SeasonDetail {
  id: number;
  season_number: number;
  name: string;
  overview?: string;
  poster_path?: string | null;
  episodes: Episode[];
}