# lib/api-zod/src/generated/types/tvDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StreamVault API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Genre } from "./genre";
import type { MediaPage } from "./mediaPage";
import type { Season } from "./season";
import type { TvDetailCredits } from "./tvDetailCredits";
import type { TvDetailVideos } from "./tvDetailVideos";

export interface TvDetail {
  id: number;
  name: string;
  overview?: string;
  poster_path?: string | null;
  backdrop_path?: string | null;
  vote_average?: number;
  vote_count?: number;
  first_air_date?: string;
  last_air_date?: string;
  number_of_seasons?: number;
  number_of_episodes?: number;
  status?: string;
  tagline?: string | null;
  genres?: Genre[];
  seasons?: Season[];
  credits?: TvDetailCredits;
  videos?: TvDetailVideos;
  similar?: MediaPage;
}